backend/pyproject.toml
[project]
name = "mini-lms-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.17.2",
    "bcrypt>=4.0.1",
    "email-validator>=2.3.0",
    "fastapi>=0.124.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]
